hetzner-deploy/requirements.txt
celery==5.4.0
redis==5.2.0
supabase==2.9.0
python-dotenv==1.0.1
finnhub-python==2.4.20
pywebpush==2.0.1
httpx==0.27.0
cryptography==44.0.0
pydantic>=2.5.0  # For FastAPI request/response models
fastapi>=0.110.0  # For chart API endpoints
uvicorn[standard]>=0.27.0  # ASGI server for FastAPI
openai>=1.0.0  # For ChartWatcher Vision API
pillow>=10.0.0  # For image processing (if needed)
reportlab>=4.0.0  # For PDF report generation (JournalBot)
